% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFUN.R
\name{logis}
\alias{logis}
\title{Evaluating Logistic Function}
\usage{
logis(x)
}
\arguments{
\item{x}{A numeric value at which the logistic function is to be evaluated.}
}
\value{
The value of the logistic function at x.
}
\description{
Evaluating the value of the logistic function at an input real number x.
}
