% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFUN.R
\name{powerxy}
\alias{powerxy}
\title{Evaluating Exponents}
\usage{
powerxy(x, y)
}
\arguments{
\item{x}{A numeric value which is to be raised to an exponent.}

\item{y}{The numeric exponent to which x needs to be raised.}
}
\value{
x^y from the input.
}
\description{
Raising one number to the power of another number.
}
